2022/day10/day10_test.go
package day10_test

import (
	"adventofcode/day10"
	"adventofcode/reader"
	"testing"

	"github.com/stretchr/testify/assert"
)

func TestSolve(t *testing.T) {
	lines := []string{
		"addx 15",
		"addx -11",
		"addx 6",
		"addx -3",
		"addx 5",
		"addx -1",
		"addx -8",
		"addx 13",
		"addx 4",
		"noop",
		"addx -1",
		"addx 5",
		"addx -1",
		"addx 5",
		"addx -1",
		"addx 5",
		"addx -1",
		"addx 5",
		"addx -1",
		"addx -35",
		"addx 1",
		"addx 24",
		"addx -19",
		"addx 1",
		"addx 16",
		"addx -11",
		"noop",
		"noop",
		"addx 21",
		"addx -15",
		"noop",
		"noop",
		"addx -3",
		"addx 9",
		"addx 1",
		"addx -3",
		"addx 8",
		"addx 1",
		"addx 5",
		"noop",
		"noop",
		"noop",
		"noop",
		"noop",
		"addx -36",
		"noop",
		"addx 1",
		"addx 7",
		"noop",
		"noop",
		"noop",
		"addx 2",
		"addx 6",
		"noop",
		"noop",
		"noop",
		"noop",
		"noop",
		"addx 1",
		"noop",
		"noop",
		"addx 7",
		"addx 1",
		"noop",
		"addx -13",
		"addx 13",
		"addx 7",
		"noop",
		"addx 1",
		"addx -33",
		"noop",
		"noop",
		"noop",
		"addx 2",
		"noop",
		"noop",
		"noop",
		"addx 8",
		"noop",
		"addx -1",
		"addx 2",
		"addx 1",
		"noop",
		"addx 17",
		"addx -9",
		"addx 1",
		"addx 1",
		"addx -3",
		"addx 11",
		"noop",
		"noop",
		"addx 1",
		"noop",
		"addx 1",
		"noop",
		"noop",
		"addx -13",
		"addx -19",
		"addx 1",
		"addx 3",
		"addx 26",
		"addx -30",
		"addx 12",
		"addx -1",
		"addx 3",
		"addx 1",
		"noop",
		"noop",
		"noop",
		"addx -9",
		"addx 18",
		"addx 1",
		"addx 2",
		"noop",
		"noop",
		"addx 9",
		"noop",
		"noop",
		"noop",
		"addx -1",
		"addx 2",
		"addx -37",
		"addx 1",
		"addx 3",
		"noop",
		"addx 15",
		"addx -21",
		"addx 22",
		"addx -6",
		"addx 1",
		"noop",
		"addx 2",
		"addx 1",
		"noop",
		"addx -10",
		"noop",
		"noop",
		"addx 20",
		"addx 1",
		"addx 2",
		"addx 2",
		"addx -6",
		"addx -11",
		"noop",
		"noop",
		"noop",
	}

	resultPart1, resultPart2 := day10.Solve(lines)
	expectedPart2 := "##..##..##..##..##..##..##..##..##..##..###...###...###...###...###...###...###.####....####....####....####....####....#####.....#####.....#####.....#####.....######......######......######......###########.......#######.......#######....."
	assert.Equal(t, 13140, resultPart1)
	assert.Equal(t, expectedPart2, resultPart2)
}

func TestSolveRealInput(t *testing.T) {
	resultPart1, resultPart2 := day10.Solve(reader.ReadInput(10))
	expectedPart2 := "###..#....###...##..####.###...##..#....#..#.#....#..#.#..#.#....#..#.#..#.#....#..#.#....#..#.#..#.###..###..#....#....###..#....###..####.#....#..#.#....#....#....#....#....#..#.#....#..#.#..#.#....#....####.#....#..#.#....###...##..####."
	assert.Equal(t, 12560, resultPart1)
	assert.Equal(t, expectedPart2, resultPart2)
}

func TestIsSignalStrengthCycle(t *testing.T) {
	assert.False(t, day10.IsSignalStrengthCycle(1))
	assert.False(t, day10.IsSignalStrengthCycle(240))

	assert.True(t, day10.IsSignalStrengthCycle(20))
	assert.True(t, day10.IsSignalStrengthCycle(60))
	assert.True(t, day10.IsSignalStrengthCycle(100))
	assert.True(t, day10.IsSignalStrengthCycle(140))
	assert.True(t, day10.IsSignalStrengthCycle(180))
	assert.True(t, day10.IsSignalStrengthCycle(220))
}
